Moved the functions in RobotContainer to a separate file
